+++ lib/rubocop/cop/lint/ensure_return.rb
@@ -26,7 +26,9 @@ module RuboCop
       #     do_something_else
       #   end
       class EnsureReturn < Cop
-        MSG = 'Do not return from an `ensure` block.'.freeze
+        MSG = 'Do not return from an `ensure` block, it changes the control ' \
+              'flow as if a `rescue Exception` clause was in place before ' \
+              'the `ensure` clause.'.freeze
 
         def on_ensure(node)
           ensure_body = node.body